Extract NotQuery into a separate file
